Enable saving videos.
Adds handling of video (i.e. timelapse) types to the repository, so that
videos can now be saved.

Also introduces support for mixed images & videos in gallery, so that videos
clicked on from a "Saved" context actually play (as opposed to just displaying
the still image, if the container is not explicitly the timelapse folder.)
Basically this moves knowledge of the type (still or video) of an image record
to the record and its handler component, instead of having that knowledge
exist only at the gallery view level.