Rewrite how media data is stored.
Usage of the app revealed that duplicate saves are a problem. The obvious
proximate fix is to make the image's handle a direct hash of its contents (so
that its ID is stable) rather than including date and path in the hash. That
is, where originally we hashed to prevent collisions and a Save operation
resulted in a copy with a new hash, we explicitly want collisions as a way to
prevent duplicate saves.

However then we have to handle the collisions.

This change moves to a single directory containing all media files, with the
various Kinds moving to "pins" of a data file via filesystem symlinks. Thus
the same image can appear in multiple categories via pins from multiple
locations, and copies are never made. The garbage collection system is updated
to match. User save operations are now idempotent (meaning they now amount to
a pin operation, which is a no-op if the image is already pinned.)

User notification of Save collisions is not yet plumbed into the UI (which
will happen with the overall error-handling pass which needs to be done
anyway.)